requirements.txt
Django==1.10.6
django-registration-redux==1.4
MySQL-python==1.2.5
olefile==0.44
Pillow==4.0.0
pkg-resources==0.0.0
psycopg2==2.6.2
python-dateutil==2.6.0
python-magic==0.4.12
six==1.10.0
django-crispy-forms==1.6.1
python-social-auth==0.3.6
django-allauth==0.31.0
django-extensions==1.7.8
Werkzeug
social-auth-app-django==1.1.0
django-facebook-connect==1.0.1
django-debug-toolbar
